Moved citnames and intercept utilities into bear as subcommands
