+++ source/bear/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Create a static library, which is used for unit tests and the final shared library.
 add_library(bear_a OBJECT)
-target_include_directories(bear_a PUBLIC source/)
+target_include_directories(bear_a PUBLIC source/ ../citnames/include/ ../intercept/include/)
 target_sources(bear_a
         PRIVATE
             source/Application.cc
@@ -12,6 +12,8 @@ target_link_libraries(bear_a PUBLIC
         sys_a
         flags_a
         fmt::fmt
+        citnames_a
+        intercept_a
         spdlog::spdlog)
 
 # Create an executable from the sub projects.
